Refactor:Server
