--- core/src/main/resources/META-INF/services/io.seata.core.store.LockStore
@@ -1,1 +1,0 @@
-io.seata.core.store.db.LockStoreDataBaseDAO
