--- core/src/main/resources/META-INF/services/io.seata.core.store.LogStore
@@ -1,1 +1,0 @@
-io.seata.core.store.db.LogStoreDataBaseDAO
