+++ server/src/main/java/io/seata/server/coordinator/DefaultCoordinator.java
@@ -224,7 +224,7 @@ public class DefaultCoordinator extends AbstractTCInboundHandler implements Tran
                     globalSession.getXid() + " " + globalSession.getStatus() + " " + globalSession.getBeginTime() + " "
                         + globalSession.getTimeout());
             }
-            boolean shouldTimeout = globalSession.lockAndExcute(() -> {
+            boolean shouldTimeout = SessionHolder.lockAndExecute(globalSession, () -> {
                 if (globalSession.getStatus() != GlobalStatus.Begin || !globalSession.isTimeout()) {
                     return false;
                 }
