+++ server/src/main/java/io/seata/server/storage/db/store/LockStoreSqls.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.core.store.db;
+package io.seata.server.storage.db.store;
 
 import io.seata.common.exception.NotSupportYetException;
 import io.seata.core.constants.DBType;
