+++ server/src/test/java/WriteStoreTest.java
@@ -143,6 +143,12 @@ public class WriteStoreTest {
                 }
 
                 @Override
+                public <T> T lockAndExecute(GlobalSession globalSession, GlobalSession.LockCallable<T> lockCallable)
+                        throws TransactionException {
+                    return null;
+                }
+
+                @Override
                 public void onBegin(GlobalSession globalSession) throws TransactionException {
 
                 }
