+++ server/src/test/java/io/seata/server/lock/file/FileLockManagerImplTest.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.server.lock.memory;
+package io.seata.server.lock.file;
 
 import io.seata.common.XID;
 import io.seata.core.model.BranchType;
