+++ server/src/test/java/io/seata/server/store/SessionStoreTest.java
@@ -25,7 +25,7 @@ import io.seata.core.model.BranchStatus;
 import io.seata.core.model.BranchType;
 import io.seata.core.model.GlobalStatus;
 import io.seata.server.lock.LockManager;
-import io.seata.server.lock.memory.FileLockManagerForTest;
+import io.seata.server.lock.file.FileLockManagerForTest;
 import io.seata.server.session.BranchSession;
 import io.seata.server.session.GlobalSession;
 import io.seata.server.session.SessionHelper;
